Trying to fix segmentation fault

+++ Code/MonoMutliViewClassifiers/utils/Dataset.py
@@ -35,14 +35,14 @@ def getValue(DATASET):
 
 def extractSubset(matrix, usedIndices):
     if sparse.issparse(matrix):
-        newIndptr = np.zeros(len(usedIndices)+1, dtype=np.int16)
+        newIndptr = np.zeros(len(usedIndices)+1, dtype=int)
         oldindptr = matrix.indptr
         print oldindptr
         for exampleIndexIndex, exampleIndex in enumerate(usedIndices):
             newIndptr[exampleIndexIndex+1] = newIndptr[exampleIndexIndex]+(oldindptr[exampleIndex+1]-oldindptr[exampleIndex])
         print newIndptr
         newData = np.ones(newIndptr[-1], dtype=bool)
-        newIndices =  np.zeros(newIndptr[-1], dtype=np.int32)
+        newIndices =  np.zeros(newIndptr[-1], dtype=int)
         oldIndices = matrix.indices
         for exampleIndexIndex, exampleIndex in enumerate(usedIndices):
             newIndices[newIndptr[exampleIndexIndex]:newIndptr[exampleIndexIndex+1]] = oldIndices[oldindptr[exampleIndex]: oldindptr[exampleIndex+1]]
